+++ controllers/index__controller.js
@@ -1,24 +1,33 @@
 import { fetchAttractionsData } from "../views/index__view--attractions.js";
 import { listBarApi } from "../views/index__view--listbar.js";
 import { searchInputApi } from "../views/index__view--search.js";
-import { getDomElements, setupEventListeners } from "../utils/user__dom.js"
+
+// utils //
+import { getUserDomElements, setupEventListeners } from "../utils/user__dom.js"
 import { detectJwt } from "../utils/user__auth.js"
+import { navigationLeftToHomePage } from "../utils/homepage.js"
 
 
 window.addEventListener("DOMContentLoaded", function(){
     fetchAttractionsData();
     listBarApi();
     searchInputApi();
+    navigationLeftToHomePage();
 
-    const elements = getDomElements();
+    // user //
+    const elements = getUserDomElements();
     setupEventListeners(elements);
     detectJwt(elements)
+
+    // booking //
+    openBookingPage()
 });
 
+const openBookingPage = () => {
+    document.querySelector(".navigation__right-booking").addEventListener("click",() => {
+    const bookingPage = `http://${window.location.host}/booking`
+    window.location.href =  bookingPage;
+    });
+};
 
-export const navigationLeftToHomePage = () => {
-    document.querySelector(".navigation__left").addEventListener("click", function() {
-    const homePage = `http://${window.location.host}`;
-    window.location.href = homePage;
-    }
-)};
+
