+++ static/booking.html
@@ -3,6 +3,8 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <script type="module" src="/controllers/booking__controller.js"></script>
+  <link rel="stylesheet" href="/views/layouts/user.css">
   <link rel="stylesheet" href="/views/layouts/booking.css">
   <title>Taipei Day Trip</title>
 </head>
