+++ views/booking__view.js
@@ -1,0 +1,4 @@
+
+export const renderBookingPage = () => {
+    console.log("render")
+}
