+++ views/index__view--attractions.js
@@ -1,5 +1,5 @@
 import { handleObserver, getNextPage } from "../controllers/index__controller--pagination.js"
-import { navigationLeftToHomePage } from "../controllers/index__controller.js"
+import { navigationLeftToHomePage } from "../utils/homepage.js"
